refactor: remove deprecated claudeNew integration and clean up related code
